# Improve performance of reference definition list parsing