--- a/lib/parser_block.mjs
+++ b/lib/parser_block.mjs
@@ -22,15 +22,15 @@ import r_paragraph from './rules_block/paragraph.mjs'
 const _rules = [
   // First 2 params - rule name & source. Secondary array - list of rules,
   // which can be terminated by this one.
-  ['table',      r_table,      ['paragraph', 'reference']],
+  ['table',      r_table,      ['paragraph']],
   ['code',       r_code],
-  ['fence',      r_fence,      ['paragraph', 'reference', 'blockquote', 'list']],
-  ['blockquote', r_blockquote, ['paragraph', 'reference', 'blockquote', 'list']],
-  ['hr',         r_hr,         ['paragraph', 'reference', 'blockquote', 'list']],
-  ['list',       r_list,       ['paragraph', 'reference', 'blockquote']],
+  ['fence',      r_fence,      ['paragraph', 'blockquote', 'list']],
+  ['blockquote', r_blockquote, ['paragraph', 'blockquote', 'list']],
+  ['hr',         r_hr,         ['paragraph', 'blockquote', 'list']],
+  ['list',       r_list,       ['paragraph', 'blockquote']],
   ['reference',  r_reference],
-  ['html_block', r_html_block, ['paragraph', 'reference', 'blockquote']],
-  ['heading',    r_heading,    ['paragraph', 'reference', 'blockquote']],
+  ['html_block', r_html_block, ['paragraph', 'blockquote']],
+  ['heading',    r_heading,    ['paragraph', 'blockquote']],
   ['lheading',   r_lheading],
   ['paragraph',  r_paragraph]
 ]