@RomanHotsiy 2024-01-29
oh, it looks like it can be improved more by removing the redundant code in the reference parsing.

I'll adjust it later today.