@rlidwka 2024-02-03
Here's a test which would have different behavior:

```
[foo]: /url 'title
 - - -
'

[foo]
```

It was `<hr>`, now it's a reference link.