@RomanHotsiy 2024-02-15
Superseded by https://github.com/markdown-it/markdown-it/pull/1004